Refactoring Exercise files structure
